requirements.txt
Django==5.2.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
reportlab==4.0.7
qrcode[pil]==7.4.2
PyPDF2==3.0.1
Pillow==10.2.0

